Updated makefile for llvm internal build

+++ llvm-build-script.sh
@@ -1,19 +1,24 @@
-cd dependencies
-tar -xzvf llvm-3.0.tar.gz
-mkdir build
+if [ ! -d "dependencies/llvm" ] ; then
+    cd dependencies
+    tar -xzvf llvm-3.0.tar.gz
+    mkdir build
 
-cd build
-../llvm-3.0.src/configure --enable-optimized --disable-assertions
-make -j
-cd ..
-mkdir llvm
-mv build/Release/* llvm
-mkdir llvm/include
-mkdir llvm/include/llvm
+    cd build
+    ../llvm-3.0.src/configure --enable-optimized --disable-assertions
+    make -j
+    cd ..
+    mkdir llvm
+    mv build/Release/* llvm
+    mkdir llvm/include
+    mkdir llvm/include/llvm
 
-mv llvm-3.0.src/include/llvm/* llvm/include/llvm/
+    mv llvm-3.0.src/include/llvm/* llvm/include/llvm/
 
-cp -rfl build/include/llvm/ llvm/include/
+    cp -rfl build/include/llvm/ llvm/include/
 
-rm -rf build
-rm -rf llvm-3.0.src
+    rm -rf build
+    rm -rf llvm-3.0.src
+else
+    echo "LLVM already configured"
+
+fi
